Added three functions to game
The getComputerChoice() function creates a random number between
0-2 and returns a string based on the number. It is used as the
computer choice in the game.

The playRound() function accepts the computerChoice and
playerSelection arguments. It compares the player selection to the
computer selection and returns a string informing the user of
victory or defeat.

The game() function runs playRound() 5 times in order to determine
the winner: best of 5.