Create new branch/add UI
Added new buttons that allows the user to play the game without
typing into a prompt. No longer need game() function. Refactored
playRound() in order to accept events and call a new function when
the player or computer reaches a score of 5.

The new endGame() updates global variables after the game is over.